+++ src/components/signUp/OwnerForm.tsx
@@ -1,86 +1,63 @@
-import { FormEvent, useEffect, useRef } from 'react';
-import signUpForm from '@/components/signup/SignUpForm.module.scss';
+import signUpForm from '@/components/signUp/SignUpForm.module.scss';
+import { MemberFormElements, MemberFormProps } from '@/components/signUp/MemberForm';
+import { FormEvent, useRef } from 'react';
 import { useUser } from '@/hooks/useUser';
 import { useNavigate } from 'react-router-dom';
 import SearchButton from '@/components/common/SeachButton';
 import PasswordInput from '@/components/common/PasswordInput';
-import { SignUpMemberData, SignUpOwnerData } from '@/types/userData';
 import Loading from '@/components/common/Loading';
 import NumberInput from '@/components/common/NumberInput';
-import { toast } from 'react-toastify';
+import { SignUpOwnerData } from '@/types/userData';
 import { formElementValueCheck, formPasswordCheck } from '@/utils/formElementValueCheck';
+import { handleToastError } from '@/utils/handleToast';
 
-export interface MemberFormElements extends HTMLFormControlsCollection {
-  email: HTMLInputElement;
-  password: HTMLInputElement;
-  passwordConfirm: HTMLInputElement;
-  phoneNumber: HTMLInputElement;
-  name: HTMLInputElement;
-  birthDate: HTMLInputElement;
-  occupation: HTMLInputElement;
-  address: HTMLInputElement;
-  addressDetail: HTMLInputElement;
-  gender: HTMLSelectElement;
+interface OwnerFormElements extends Omit<MemberFormElements, 'occupation'> {}
+
+interface OwnerForm extends HTMLFormElement {
+  readonly elements: OwnerFormElements;
 }
 
-interface MemberForm extends HTMLFormElement {
-  readonly elements: MemberFormElements;
-}
+export interface OwnerFormProps extends MemberFormProps {}
 
-export interface MemberFormProps {
-  sendSignUpData: ({ data, isOwner }: { data: SignUpMemberData | SignUpOwnerData; isOwner: boolean }) => void;
-  openPostModal: () => void;
-  roadAddress: string;
-}
-
-const MemberForm = ({ sendSignUpData, openPostModal, roadAddress }: MemberFormProps) => {
+const OwnerForm = ({ sendSignUpData, openPostModal, roadAddress }: OwnerFormProps) => {
   const formRef = useRef<HTMLFormElement>(null);
   const { data: userData, isLoading } = useUser();
   const navigate = useNavigate();
 
-  const onSubmit = (e: FormEvent<MemberForm>) => {
+  const onSubmit = (e: FormEvent<OwnerForm>) => {
     try {
       e.preventDefault();
       const currentTarget = e.currentTarget;
-      if (!currentTarget) return;
+      if (!e.currentTarget) return;
 
-      const { email, name, phoneNumber, birthDate, gender, occupation, addressDetail, password, passwordConfirm } =
-        currentTarget.elements;
+      const { email, password, passwordConfirm, name, phoneNumber, birthDate, gender, addressDetail } =
+        e.currentTarget.elements;
 
       formPasswordCheck({ currentTarget, password: password.value, passwordConfirm: passwordConfirm.value });
 
       const address = (roadAddress + ' ' + addressDetail.value).trim();
 
-      const data: SignUpMemberData = {
+      const data: SignUpOwnerData = {
         email: email.value,
         password: password.value,
-        passwordConfirm: passwordConfirm.value, //삭제 예정
+        passwordConfirm: passwordConfirm.value, //삭제예정
         name: name.value,
         phoneNumber: phoneNumber.value,
-        address,
-        gender: gender.value,
+        businessRegistrationNumber: 'string', //삭제예정
         birthDate: birthDate.value,
-        occupation: occupation.value,
+        gender: gender.value,
+        address,
       };
 
-      formElementValueCheck<MemberForm, SignUpMemberData>({ currentTarget, data });
+      formElementValueCheck<OwnerForm, SignUpOwnerData>({ currentTarget: e.currentTarget, data });
       sendSignUpData({
-        isOwner: false,
+        isOwner: true,
         data,
       });
     } catch (error) {
-      console.error(error);
-      if (error instanceof Error) {
-        toast.error(error.message);
-      }
+      handleToastError(error);
     }
   };
-
-  // useEffect(() => {
-  //   if (isLoading === false && !userData?.email) {
-  //     navigate('/signin');
-  //   }
-  // }, [isLoading, userData]);
 
   if (isLoading) {
     return <Loading style={{ height: '42.5rem' }} />;
@@ -89,18 +66,14 @@ const MemberForm = ({ sendSignUpData, openPostModal, roadAddress }: MemberFormPr
   return (
     <form ref={formRef} onSubmit={onSubmit} className={signUpForm['form']}>
       {/* email readOnly */}
-      <label htmlFor="email" className={signUpForm['email']}>
-        이메일
-      </label>
-      <input name="email" type="email" tabIndex={-1} />
+      <label htmlFor="email">이메일</label>
+      <input className={signUpForm['email']} name="email" type="email" tabIndex={-1} />
       <PasswordInput passwordName="password" passwordConfirmName="passwordConfirm" />
       {/* name phone */}
       <label htmlFor="name">이름</label>
       <input name="name" type="text" />
       <label htmlFor="phoneNumber">연락처</label>
       <NumberInput name="phoneNumber" maxLength={11} pattern={/(^\d{3})(\d{3,4})(\d{4}$)/} />
-      <label htmlFor="occupation">직업</label>
-      <input name="occupation" type="text" />
       {/* address */}
       <label htmlFor="address">주소</label>
       <div className={signUpForm['address']}>
@@ -136,4 +109,4 @@ const MemberForm = ({ sendSignUpData, openPostModal, roadAddress }: MemberFormPr
   );
 };
 
-export default MemberForm;
+export default OwnerForm;
